+++ noodles-cram/CHANGELOG.md
@@ -14,7 +14,8 @@
 
 ### Changed
 
-  * cram/crai/record: Change alignment start to an `Option<Position>`.
+  * cram/crai/record: Change alignment start to an `Option<Position>` and
+    alignment span to a `usize`.
 
   * cram/record: Move `mapping_quality`, `read_name`, and `quality_scores` to
     the implementation of `sam::AlignmentReader`.